requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
sqlparse==0.4.4
typing_extensions==4.7.1
